Added confirmation to delete requests. Added cancel buttons

+++ public/games.js
@@ -13,15 +13,18 @@ function reviewCheck(game_ID, console_ID, game_name, console_name) {
 };
 
 function deleteGame(console_ID, game_ID) {
-    $.ajax({
-        url: "/games/deleteGame?console_ID=" + console_ID + "&game_ID=" + game_ID,
-        type: "DELETE",
-        success: function () {
-            alert("game deleted");
-            window.location.replace("/");
-        },
-        error: function (results) {
-            alert(results.statusText);
-        }
-    })
+    var confirmation = confirm("delete game?");
+    if (confirmation) {
+        $.ajax({
+            url: "/games/deleteGame?console_ID=" + console_ID + "&game_ID=" + game_ID,
+            type: "DELETE",
+            success: function () {
+                alert("game deleted");
+                window.location.replace("/");
+            },
+            error: function (results) {
+                alert(results.statusText);
+            }
+        })
+    }
 };
