+++ public/profile.js
@@ -18,15 +18,18 @@ $(document).ready(function () {
 });
 
 function deleteReview(review_ID) {
-    $.ajax({
-        url: "/profile/deleteReview?review_ID=" + review_ID,
-        type: "DELETE",
-        success: function () {
-            alert("review deleted");
-            window.location.reload(true);
-        },
-        error: function (results) {
-            alert(results.statusText);
-        }
-    })
+    var confirmation = confirm("delete review?");
+    if (confirmation) {
+        $.ajax({
+            url: "/profile/deleteReview?review_ID=" + review_ID,
+            type: "DELETE",
+            success: function () {
+                alert("review deleted");
+                window.location.reload(true);
+            },
+            error: function (results) {
+                alert(results.statusText);
+            }
+        })
+    }
 };
